Value.String(): Handle unrecognized types better.
If you were to, for example, create a value of type GstNvCamFlickerMode --
in other words, something not an integer or a string -- and then attempt to
print the value via Printf("%v"), you would end up with a recursive loop:
	- Printf() looks for a String() method for Value,
	- Printf() calls glib.Value.String(),
	- Which calls glib.Value.Get(),
	- Which doesn't recognize the type, so returns the Value argument
	  unchanged.
	- Go notices the returned value isn't a string, so tries to convert
	  it to one by calling glib.Value.String() again.
	- Rinse, recurse...

Break this loop in Value.String() by checking if the value we got back from
Value.Get() is the same used to invoke us.  If so, return a string saying we
don't recognize the type ("unknown value type...").  This is needed to
prevent gst.DeserializeValue() from disappearing down a rabbit hole.

Pursuant to EM-559.